Add files via upload
Voice assistants are programs on digital devices that listen and respond to verbal commands. A user can say, “What’s the 
weather?” and the voice assistant will answer with the weather report for that day and location. They could say, “Tell me a joke,” and 
the assistant will jump into a tale. The user could even say, “play music,” and song will be played ! Voice assistants are so easy to use 
that many people forget to stop and WONDER how they work. How do voice assistants understand us? Is it magic? A complex system 
of codes? An actual person listening on the other end? The answer is less complicated than you might think. The application works like 
Siri, Google Assistant etc. The U.I of the application is self explainable and very minimum. It takes voice as input. The system is being 
designed in such a way that all the services provided by the mobile devices are accessible by the end user on the user's voice command.


Download TESTBANSI.py file ,and run on any compiler with Python extension.